--- src/ConnectionManager.ts
@@ -1,8 +1,0 @@
-import * as isOnline from 'is-online'
-
-export class ConnectionManager{
-
-    static thereIsInternetConnection(){
-        return isOnline();
-    }
-}
