+++ src/components/readers/offlineReaderRender/PdfRender.ts
