+++ src/services/ConectivityService.ts
