--- test/GoogleApiResponse.ts
@@ -1,80 +1,0 @@
-/* tslint-disable */
-export default {
-  kind: "books#volumes",
-  totalItems: 62,
-  // tslint:disable-next-line:object-literal-sort-keys
-  items: [
-    {
-      id: "g06dIAAAAEAJ",
-      kind: "books#volume",
-      // tslint:disable-next-line:object-literal-sort-keys
-      etag: "p+Wl7yoprvI",
-      selfLink: "https://www.googleapis.com/books/v1/volumes/g06dIAAAAEAJ",
-      volumeInfo: {
-        title: "Winning the Loser's Game",
-        // tslint:disable-next-line:object-literal-sort-keys
-        authors: [
-          "Charles D. Ellis",
-        ],
-        publisher: "McGraw-Hill Education",
-        // tslint:disable-next-line:max-line-length
-        description: "Everything you need to know for long-term stock-market success—with critical new information about indexing, ETFs, and 401(k)s From one of today's most respected authorities on stock investing, Winning the Loser's Game explains how you ...",
-        readingModes: {
-          text: true,
-          // tslint:disable-next-line:object-literal-sort-keys
-          image: false,
-        },
-        printType: "BOOK",
-        maturityRating: "NOT_MATURE",
-        allowAnonLogging: false,
-        contentVersion: "full-1.0.0",
-        imageLinks: {
-          // tslint:disable-next-line:max-line-length
-          smallThumbnail: "http://books.google.com/books/content?id=g06dIAAAAEAJ&printsec=frontcover&img=1&zoom=5&uvs=3&source=gbs_api",
-          // tslint:disable-next-line:max-line-length
-          thumbnail: "http://books.google.com/books/content?id=g06dIAAAAEAJ&printsec=frontcover&img=1&zoom=1&uvs=3&source=gbs_api",
-        },
-        previewLink: "http://books.google.co.ve/books?id=g06dIAAAAEAJ&printsec=frontcover&hl=&source=gbs_api",
-        infoLink: "https://play.google.com/store/books/details?id=g06dIAAAAEAJ&source=gbs_api",
-        canonicalVolumeLink: "https://market.android.com/details?id=book-g06dIAAAAEAJ",
-      },
-      userInfo: {
-        copy: {
-          limitType: "unlimited",
-        },
-        isPurchased: true,
-        isUploaded: true,
-        // tslint:disable-next-line:object-literal-sort-keys
-        isInMyBooks: true,
-        updated: "2018-05-10T21:49:47.748Z",
-        acquiredTime: "2018-05-10T21:49:47.748Z",
-        userUploadedVolumeInfo: {
-          processingState: "COMPLETED_SUCCESS",
-        },
-        entitlementType: 4,
-      },
-      saleInfo: {
-        country: "VE",
-        saleability: "NOT_FOR_SALE",
-        // tslint:disable-next-line:object-literal-sort-keys
-        isEbook: true,
-      },
-      accessInfo: {
-        country: "VE",
-        viewability: "UNKNOWN",
-        // tslint:disable-next-line:object-literal-sort-keys
-        embeddable: false,
-        publicDomain: false,
-        textToSpeechPermission: "ALLOWED",
-        epub: {
-          // tslint:disable-next-line:max-line-length
-          downloadLink: "http://books.google.co.ve/books/download/Winning_the_Loser_s_Game_7th_Edition_201?id=g06dIAAAAEAJ&hl=&output=uploaded_content&source=gbs_api",
-          isAvailable: true,
-        },
-        webReaderLink: "http://play.google.com/books/reader?id=g06dIAAAAEAJ&hl=&printsec=frontcover&source=gbs_api",
-        accessViewStatus: "FULL_PURCHASED",
-        quoteSharingAllowed: false,
-      },
-    },
-  ],
-};
